Mark fsevents as optional
It is a transitive dependency introduced via `mocha` (not currently used
but may be used later), so we can't prevent it from included in
package-lock.json.

`chokidar` (directly depends on `fsevent`) optionally depends on it.

By removing package-lock.json and regenerating it with
`npm i --lockfile-version 1`, `fsevents` was marked optional. To
minimize diff of this commit, the file was reverted and then marked
optional manually.